Move separate_union_literals to typeopts.py; fix enum literal printing
